Removed xxhash32 in favour of crc32, fixed maphash and fndatahandler to use proper crc32 hashing

+++ game/server/svglobals.cpp
@@ -134,7 +134,7 @@ static bool IsVerifiedMap()
 	{
 		char mapfile[MAX_PATH];
 		_snprintf(mapfile, sizeof(mapfile), "%s/maps/%s.bsp", MSGlobals::AbsGamePath.c_str(), MSGlobals::MapName.c_str());
-		unsigned int hash = GetFileCheckSum(mapfile);
+		unsigned long hash = GetFileCheckSum(mapfile);
 		if (!FnDataHandler::IsVerifiedMap(MSGlobals::MapName.c_str(), hash))
 			return false;
 	}
